Added code for storing a list of faces, with pointers to neighbouring cells

+++ include/examples/mesh_examples.hpp
@@ -9,4 +9,4 @@ using namespace minicombust::geometry;
 Mesh<double> *load_boundary_box_mesh(double box_size);
 
 // <typename T>
-Mesh<double> *load_global_mesh(double mesh_dim, int elements_per_dim);
+Mesh<double> *load_global_mesh(double mesh_dim, uint64_t elements_per_dim);
